feature: implemented tree method
